requirements.txt
transformers
pillow
huggingface_hub==0.23.5
pymongo
accelerate>=0.26.0